+++ include/BDSCollimatorJaw.hh
@@ -19,41 +19,75 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef BDSCOLLIMATORJAW_H
 #define BDSCOLLIMATORJAW_H
 
-#include "BDSCollimator.hh"
+#include "BDSAcceleratorComponent.hh"
+
+#include "globals.hh" // geant4 types / globals
+#include "G4Material.hh"
 
 class G4Colour;
+class G4VSolid;
 
 /**
- * @brief A class for a rectangular collimator.
+ * @brief Collimator with only two jaw and no top bit.
+ *
+ * @author Laurie Nevay
  */
 
-class BDSCollimatorJaw: public BDSCollimator
+class BDSCollimatorJaw: public BDSAcceleratorComponent
 {
 public:
-  BDSCollimatorJaw(G4String  name,
-		 G4double  length,
-		 G4double  horizontalWidth,
-		 G4double  xApertureIn        = 0,
-		 G4double  yApertureIn        = 0,
-		 G4double  xOutApertureIn     = 0,
-		 G4double  yOutApertureIn     = 0,
-		 G4double  jaw1Offset         = 0,
-		 G4double  jaw2Offset         = 0,
-		 G4String  collimatorMaterial = "copper",
-		 G4String  vacuumMaterial     = "vacuum",
-		 G4Colour* colour             = nullptr);
-  virtual ~BDSCollimatorJaw(){;};
+  BDSCollimatorJaw(G4String  nameIn,
+		   G4double  lengthIn,
+		   G4double  horizontalWidthIn,
+		   G4double  xApertureIn,
+		   G4double  yApertureIn,
+		   G4double  jaw1OffsetIn,
+		   G4double  jaw2OffsetIn,
+		   G4double  xHalfGapIn,
+		   G4double  yHalfHeightIn,
+		   G4bool    buildLeftJawIn,
+		   G4bool    buildRightJawIn,
+		   G4Material* collimatorMaterialIn,
+		   G4Material* vacuumMaterialIn,
+		   G4Colour* colourIn             = nullptr);
+  virtual ~BDSCollimatorJaw();
 
-  virtual void BuildInnerCollimator();
+  /// Accessor for collimator material.
+  virtual G4String Material() const {return collimatorMaterial->GetName();}
+
+protected:
+  virtual void Build();
+  
+  virtual void BuildContainerLogicalVolume();
+
+  ///@{ Geometrical objects:
+  G4VSolid* jawSolid;
+  ///@}
+
+  G4double    horizontalWidth;    ///< Horizontal width.
+  G4double    xAperture;          ///< Aperture at entrance in x dimension
+  G4double    yAperture;          ///< Aperture at entrance in y dimension
+  G4double    jaw1Offset;         ///< Offset of jaw 1
+  G4double    jaw2Offset;         ///< Offset of jaw 2
+  G4double    jcolAperture;       ///< Jaw collimator aperture width
+  G4double    apertureIsVertical; ///< Jaw collimator orientation
+  G4double    xHalfGap;           ///< Half gap separation between jaws.
+  G4double    jawHalfWidth;       ///< Half width of each jaw.
+  G4double    yHalfHeight;        ///< Half height of each jaw.
+  G4bool      buildLeftJaw;       ///< Build left jaw or not.
+  G4bool      buildRightJaw;      ///< Build right jaw or not.
+  G4Material* collimatorMaterial; ///< Material
+  G4Material* vacuumMaterial;     ///< Vacuum material
+  G4Colour*   colour;             ///< Colour of collimator
 
 private:
   /// Private default constructor to force the use of the supplied one.
-  BDSCollimatorJaw();
+  BDSCollimatorJaw() = delete;
 
   /// @{ Assignment and copy constructor not implemented nor used
   BDSCollimatorJaw& operator=(const BDSCollimatorJaw&) = delete;
   BDSCollimatorJaw(BDSCollimatorJaw&) = delete;
-  ///@}
+  /// @}
 };
 
 #endif